requirements.txt
python-dateutil==2.8.2
requests==2.28.2
python-dotenv==1.0.0
terminaltables==3.1.10